debug/requirements.txt
# SecureWatch debug tooling dependencies
# Install: pip install -r requirements.txt

httpx>=0.27.0          # async HTTP client (replay_runner)
python-dotenv>=1.0.0   # .env file loading
pytest>=8.0.0          # test runner (contract_tests)
pytest-asyncio>=0.23.0 # async pytest support
supabase>=2.4.0        # Supabase Python client
